requirements.txt
instructor[google-generativeai, vertexai]
fastapi[standard]
python-dotenv
uvicorn
redis